+++ app/javascript/vue/tasks/collection_objects/filter/app.vue
@@ -60,37 +60,50 @@
             <span class="separate-left separate-right">|</span>
             <csv-button
               :url="urlRequest"
-              :options="{ fields: csvFields }"/>
+              :options="{ fields: csvFields }"
+            />
             <dwc-download
               class="margin-small-left"
               :params="$refs.filterComponent.parseParams"
-              :total="pagination.total"/>
+              :total="pagination.total"
+            />
             <dwc-reindex
               class="margin-small-left"
               :params="$refs.filterComponent.parseParams"
               :total="pagination.total"
             />
+            <match-button
+              :ids="ids"
+              :url="urlRequest"
+              class="margin-small-left"
+            />
           </div>
         </div>
         <div
+          v-if="pagination"
           class="flex-separate margin-medium-bottom"
-          v-if="pagination">
+        >
           <pagination-component
             v-if="pagination"
-            @nextPage="loadPage"
-            :pagination="pagination"/>
+            @next-page="loadPage"
+            :pagination="pagination"
+          />
           <pagination-count
             :pagination="pagination"
-            v-model="per"/>
+            v-model="per"
+          />
         </div>
         <list-component
           v-if="Object.keys(list).length"
           v-model="ids"
           :list="list"
-          @onSort="list.data = $event"/>
+          @on-sort="list.data = $event"
+        />
         <h2
           v-if="alreadySearch && !list"
-          class="subtle middle horizontal-center-content no-found-message">No records found.
+          class="subtle middle horizontal-center-content no-found-message"
+        >
+          No records found.
         </h2>
       </div>
     </div>
@@ -108,8 +121,11 @@ import GetPagination from 'helpers/getPagination'
 import DwcDownload from './components/dwcDownload.vue'
 import DwcReindex from './components/dwcReindex.vue'
 import SelectAll from './components/selectAll.vue'
+import MatchButton from './components/matchButton.vue'
 
 export default {
+  name: 'FilterCollectionObjects',
+
   components: {
     PaginationComponent,
     FilterComponent,
@@ -118,7 +134,8 @@ export default {
     PaginationCount,
     DwcDownload,
     DwcReindex,
-    SelectAll
+    SelectAll,
+    MatchButton
   },
 
   computed: {